model got updates

+++ Model.py
@@ -9,9 +9,9 @@ class model(BaseModel):
     property_id:str
     city:str
     commercial_type:str
-    size_sqm:int
-    annual_rent:int
+    size_sqm:float
+    annual_rent:float
     occupancy_status:str
-    lease_term_years:int
+    lease_term_years:float
